Refactor code and move unnecessary things.
