+++ data/students.json
@@ -1,1 +1,1 @@
-{"0":{"id":323,"name":"Mark","age":15,"parent_permission":true,"classroom":"Maths"},"1":{"id":9924,"name":"Sofia","age":18,"parent_permission":true,"classroom":"Physics"},"2":{"id":888,"name":"James","age":14,"parent_permission":true,"classroom":"Maths"}}
+{"0":{"id":323,"name":"Mark","age":15,"parent_permission":true,"classroom":"Maths"},"1":{"id":9924,"name":"Sofia","age":18,"parent_permission":true,"classroom":"Physics"},"2":{"id":888,"name":"James","age":14,"parent_permission":true,"classroom":"Maths"},"6":{"id":4086,"name":"Ammar","age":24,"parent_permission":true,"classroom":"AI"}}
