+++ data/teachers.json
@@ -1,1 +1,1 @@
-{"3":{"id":1284,"name":"Jack","age":40,"specialization":"Maths"},"4":{"id":8241,"name":"George","age":65,"specialization":"Physics"},"5":{"id":4539,"name":"Jane","age":31,"specialization":"Electronics"}}
+{"3":{"id":1284,"name":"Jack","age":40,"specialization":"Maths"},"4":{"id":8241,"name":"George","age":65,"specialization":"Physics"},"5":{"id":4539,"name":"Jane","age":31,"specialization":"Electronics"},"7":{"id":8806,"name":"Ammar","age":24,"specialization":"Software Engineering"}}
